Fix event macro
